Refactor Projection component props and interface to add a generating boolean and change download link name.
This commit refactors the `Projection` component's props to include a `generating` boolean property and changes the download link name. The `Projection` interface in the `projection.d.ts` file has been updated to extend the `Status` interface and include an optional `id` string property along with an `img` property. A `generating` property has also been added to the `Status` interface.