feat: Add useAuthorization hook
This commit adds a new hook called useAuthorization, which checks if a user is authenticated or not by making a request to the server for a cookie. If the cookie is valid, it'll return the user object and store it in the Redux store. This hook is implemented to protect the home and the community routes by dispatching login hook when a user is authenticated and logout hook when they're not.