Refactor prompt functions to use Redux for state management
This commit introduces changes to refactor prompt functions to use Redux for state management instead of `useState`. This includes the use of the `useDispatch` hook to dispatch actions and update the store using `setState` for the `setPrompt` and `setStatus` functions. Additionally, new functions for publishing and generating a prompt and a surprise me prompt have been added using the new Redux state management. The `generatePrompt`, `publishPrompt`, and `surprisemePrompt` functions are implemented as `async` functions that use `axios` to make appropriate AJAX requests.