Improve Form input field appearance
The changes in this commit add a label and icon to enhance the style of the input fields in the Form component. The label will show the icon and the input will have a padding on the left that accommodates the icon. This improves the usability of the Form component by making it more visually appealing, resulting in a better user experience.